+++ lib/auction_fun_core/entities/user.rb
@@ -5,6 +5,14 @@ module AuctionFunCore
     # User Relations class. This return simple objects with attribute readers
     # to represent data in your user.
     class User < ROM::Struct
+      def active?
+        active
+      end
+
+      def inactive?
+        !active
+      end
+
       def info
         attributes.except(:password_digest)
       end